chore(PlaceholderParser): move to 'util/application' package
